infra/requirements.txt
pydantic >= 2.0
aws-cdk-lib
constructs
boto3
boto3-stubs[mypy-boto3-dynamodb]
